fix wcf1_package.packageDir silently
This change tries to fix the `is missing or does not implement 'wcf\system\application\IApplication'` if the folder of an has been moved, but the admin forgot to update the `wcf1_package`-table. The column `packageDir` will be fixed and the config file will be rewritten automatically. This works only of the app is placed in a sub-folder of the WSC itself.