Parse smiley codes only once during the request lifetime
The parsed smilies are already semi-immutable during the request
lifetime and especially the ordering by lenght via `mb_strlen()` is an
expensive operation. Using a static runtime cache improves the
performance when rebuilding messages in bulk, reducing the time by up to
40%.

We should backport this to 3.0 once it has proven to be stable,
especially changing the `$smilies` property into a static variable can
break derived classes.