Change the DefaultAvatar’s size to 128x128 pixels
Change the DefaultAvatar’s size to a more reasonable size of 128x128 pixels.
Throughout the Suite Core avatars are usually embedded via the `getImageTag` function
but when using the direct avatar URL a larger image is usually wanted.

Firefox for example does not upscale images used in `window.Notification`
thus the avatar would be rendered as a tiny 16x16 image in this case.
This change allows to downscale accordingly in this case.