Fix logic in CronjobScheduler::resetFailedCronjobs()
This adds the missing check for `state <> READY` to the query.

Consider the following situation in a low traffic community. The
cronjob is assumed to be executed every five minutes.

0000 Cronjob is executed regularly (next = 0005, afterNext = 0010).
0012 A visitor triggers the cronjobs, afterNext is in the past, thus
     the cronjob is assumed to have failed (next = 0015,
                                            afterNext = 0020).
0027 A visitor triggers the cronjobs, afterNext is in the past, thus
     the cronjob is assumed to have failed (next = 0030,
                                            afterNext = 0035).
0039 A visitor triggers the cronjobs, afterNext is in the past, thus
     the cronjob is assumed to have failed (next = 0040,
                                            afterNext = 0045).
0100 A visitor triggers the cronjobs, afterNext is in the past, thus
     the cronjob is assumed to have failed. The cronjob now missed
     it's execution window for 3 times and is being disabled.

With this change the state is checked to not be READY, in this case
it should have been already processed in another request when
afterNextExec lies in the past:

If the state is EXECUTING then the cronjob is the actual cause of
failure, if the state is PENDING then the cronjob is innocent and
just happens to be a bystander.