Fix form field dependency check for nested dependencies
If field C depends on field B which in turn depends on field A and if field B is unavailable because of its dependencies on field A, field C's explicit dependencies on field B might have still be fulfilled even though field B is unavailable.
Instead of just checking the dependency, we must also check if the depending field fulfills its dependencies.

See #2509