Suppress the draft restore if the stored version appears to be equivalent
This does not take formatting or non-text content (such as links) into consideration, these changes will not trigger the restore overlay. We cannot reliably tell if such changes are an actual difference, e. g. `<b>` was converted to `<strong>` on the server-side, therefore this change tries to go with an balanced approach.

Closes #2602