Suppress additional native browser title of smilies in editor tab
Prior to this change, both the `a` element and the `img` element have a `title` attribute with the `a` element being a JavaScript tooltip element. This caused a JavaScript tooltip and a native browser tooltip to be shown when hovering over a smiley in the smiley selection below the editor.

See #2220