Promote a group to be the owner group
We cannot reliably determine the owner group during the upgrade, therefore it is up to an administrator to make the choice.

The notice is designed to be annoying and is present on all pages, however it does not prevent the user from carrying out tasks, e. g. setting up a proper group before promoting it to be the owner group.