added events for hijacking the notification system
This might be needed for integrations of other notification-systems/providers or marking notifications as read in an application for mobile devices without re-fetching the whole notification-list.
Currently it's not possible to handle confirmed or removed notifications correctly.
Performance should not be reduced (a lot) since those positions are not called for every request on the site.

Closes #3016