Fix handling of default box limit/sorting if not explicitly set
If a box is created via the box package installation plugin and if no limit, sort field, or sort order is given, the default value should be used instead of leaving the values `null`.

Is also fixed the problem that simply submitting the box edit form without change for a box supporting a limit but not actually having a limit causes an error message because the limit field is empty.