Move the mobile reaction button back into the menu
The separate reaction button was a nice touch, but since it is generated on the fly, this caused the elements to increase their height. Eventually, browsers would to jump to anchors on load, but the reaction buttons would mess up the scroll location.

We should think about redesigning the mobile message view and start displaying at least some buttons. However, this could potentially cause incompatibilities with styles, therefore we need to postpone this to at least 5.3.