Add Facebook Video Media Provider
The resulting HTML intentionally does not use an `<a>` element for the
fallback link, because it will prevent the SDK from embedding the video.

The SDK generally acts as a blackbox with regard to video embeds, the only
thing we can do is asking it nicely to re-parse the DOM for new stuff to
embed and hope for the best. Specifically embedding doesn't work for newly
posted contents that are loaded via Ajax. Not even if the re-scan is
scheduled for the next run of the event loop by using `setTimeout` with a
delay of `0`. For whatever reason there might be.

Resolves #3242