Take article age into account for related article list
Instead of just ordering the related articles by the number of matching
tags the articles are now also ordered by their age as a tie-breaker,
referring more recent articles. Previously the order was undefined,
usually preferring older articles, as their rows appear earlier within
the `tag_to_object` table.

It was verified that the new query does not result in a significantly
worse query plan than the old one. The only difference is the addition
of 2 `eq_ref` joins, which is the best possible join type. However we
could get rid of a separate query to pull the articleIDs, because they
are readily available by the pure virtue of having to join the article
table to get the time.

Another possible solution would be ordering by `ABS(article.time - ?)`
with the currently viewed article's time in place of the question mark.
This would prefer articles written within a similar timeframe, without
resulting in a worse query plan within my tests.

Resolves #3285