Prevent rich metacode being dysfunctional in messages just created by replying
The `isset` strategy does not work properly, because in some cases the HTML is evaluated before being placed in the final template, causing the variable to be set already.

This changes modifies the behavior to always attempt to initialize the JavaScript, but made changes to ensure that each subsequent call is idempotent.