Fix selection of related articles for old MySQL versions
MySQL < 5.7 and apparently MariaDB up to at least 10.3 is unable to deduce
that there can only be exactly a single row from the `article` table, because
it is functionally dependent on `tag_to_object.objectID`.

Help them out by using `MAX` which effectively does nothing, because we're
taking the `MAX` of a single item.

Fixes #3528.