Remove session_validate_user_agent option
This option will not play well with long-lived user sessions and the security
benefit is roughly nil. After all almost anyone is using Google Chrome on
either Windows or Android.

Additionally since the session ID is no longer embedded within the URL losing
the session ID to a 3rd party is hard.