Remove obsolete call to getSessionByUserID in SessionHandler::create()
This call will always return `null`, since the user always is a guest at that
point, passing a `null` to the database query selecting rows `WHERE userID = ?`.

Incidentally it took me a while to understand why this would not return
existing guest sessions, until I remembered that `NULL` does not equal anything
in SQL and instead `IS NULL` needs to be used. It's good that this slightly
misleading part of the code is gone.