Replace the selector for the font-family in email_html.tpl
The Emogrifier version used in 5.2 does not fully support all features
of the `:not()` selector. Depending on the environment the use of
`:not()` results in slightly different output. Specifically the number
of `:not()`s chained there caused the result to have a very large
specificity, possibly overriding the font-size of the headline tags.

Use `body, body *` instead, it will not put any useless CSS onto heading
tags and still catch all relevant elements. The biggest difference is
that the `<br>` tag will carry a `font-family` and `font-size`.