Stop filling in the user's profile on social registration
The behavior of this was very inconsistent across providers, required a custom
implementation for each of those and over the years less and less data was
provided (most notably Facebook drastically limited the available fields).

The benefit of filling in the gender and avatar here and there is almost nil,
nothing of real value is lost when removing this and it keeps the code simpler.