Download Google fonts before validating the individual SCSS
The individual SCSS will compile the style and this compiled CSS will later be
used for the style itself.

When first downloading a new Google font the font will not yet be available
during this test, thus the necessary CSS will not be included which in turn
will cause the font to fail to display.