Revert "Prevent rewriting filenames for processed files"
This change is bogus, because it completely changes StyleAction's API and it
introduces bugs (namely not updating the database). One of these was fixed in
88b962c25b4c2dd0c312c6f7e4c11f15a1263cf4, but other fields are also affected.

As this commit doesn't fix an actual bug, but only a small inelegance (namely
the logo's path changing whenever the style is changed, even if the logo is
unchanged) this can be safely reverted until a proper fix is developed.

This reverts commit cc17f302902fb62c78b4411dadc3243913f18c74.