Skip usersOnly permissions for the guest group during permission merge
The value of usersOnly permissions cannot be configured for the guest group
within the ACP, but the stored default value is taken into account during the
permission merge in UserGroupPermissionCacheBuilder nonetheless. This might result
in unexpected permissions for accounts that are not yet activated.

As an example taking away the “Can use conversations” for all groups might
still allow accounts that are not yet activated to use the conversations,
because of this preserved default.

Fix this issue by simply skipping the value of the guest group during the
merge, thus ignoring whatever value may be stored. This is safe, because the
resulting option list always includes the “Everyone” group.

see https://community.woltlab.com/thread/289879/