Ensure the style logo is properly preserved in the database when not performing changes
This bug was introduced recently when the `!$file->isProcessed()` condition was
added. It was verified that the logo's filename remained stable within the file
system. However it was not verified that it was properly stored within the
database. As the previous style variables are deleted during the update we must
explicitly set the corresponding variable, otherwise it will be empty and the
logo will not be found.

see c72c06dc7bb21c5098613ab8fafc8596c8f6ce41
see #4240