Deprecate StringUtil::getHash()
This method was a simple alias to `\sha1()` and also defined as such. Users of
this method relied on the output of this method not changing, making it
entirely useless, because there is no benefit to use this method over `sha1()`
directly.

Deprecate `StringUtil::getHash()` and update the users to use `sha1()`.