Enable Unicode support for SCSS Compiler
The large performance difference measured back when this was introduced appears
to be gone. On PHP 8 both encodings appear to be equally fast. On PHP 7.4 the
UTF-8 version is ~25% slower, which is certainly not ideal, but not as bad as a
7× increase. In any case the UTF-8 version completes in way less than 2 seconds
which I consider to be the line that the CSS compiler should not cross.

Resolves #4339
see #4274