Register reauthentication after MFA in ACP only
It was not desired that a login within the frontend also registers a
reauthentication, as the actions requiring a reauthentication are very rarely
used and can benefit from this additional password check. In the ACP however
each request requires a reauthentication, thus it makes sense to register it
during login.

see 4624ceeda0e19109245d19e883dc6fd388e95519