Improve performance of SystemCheckPage::validateWritableDirectories()
The previous implementation resulted in a syscall hell for large attachment or
image directories. RecursiveDirectoryIterator requires two syscalls for
directory entry just for iteration and that does not yet include any custom
logic.

Replace the implementation by a simple `glob()`. This is a small behavior
change, as a `*` will only check the files immediately below the directory in
question instead of recursing all the way down. However this likely is what was
intended anyway.