Use `data:` URIs to load assets in WCFSetup
With this change the HTML responses generated by WCFSetup are completely
self-contained and do not rely on external requests.

This avoids the proxying of the assets using PHP from the temporary folder at
the expense of slightly bloating the HTML with the embedded CSS.

The total size of the HTML is 530kB before gzip and 200kB after gzip.

Resolves #4394